Reorganise tests into same namespace
