+++ tests/BankAccountNormalizer/DefaultNormalizerTest.php
@@ -1,10 +1,9 @@
 <?php
 
-namespace Cs278\BankModulus\Tests;
+namespace Cs278\BankModulus\BankAccountNormalizer;
 
 use Cs278\BankModulus\BankAccount;
 use Cs278\BankModulus\BankAccountNormalized;
-use Cs278\BankModulus\BankAccountNormalizer\DefaultNormalizer;
 
 final class DefaultNormalizerTest extends \PHPUnit_Framework_TestCase
 {
