+++ tests/BankAccountNormalizer/SevenDigitNormalizerTest.php
@@ -1,9 +1,8 @@
 <?php
 
-namespace Cs278\BankModulus\Tests;
+namespace Cs278\BankModulus\BankAccountNormalizer;
 
 use Cs278\BankModulus\BankAccount;
-use Cs278\BankModulus\BankAccountNormalizer\SevenDigitNormalizer;
 
 final class SevenDigitNormalizerTest extends \PHPUnit_Framework_TestCase
 {
