+++ tests/BankAccountNormalizer/SixDigitNormalizerTest.php
@@ -1,9 +1,8 @@
 <?php
 
-namespace Cs278\BankModulus\Tests;
+namespace Cs278\BankModulus\BankAccountNormalizer;
 
 use Cs278\BankModulus\BankAccount;
-use Cs278\BankModulus\BankAccountNormalizer\SixDigitNormalizer;
 
 final class SixDigitNormalizerTest extends \PHPUnit_Framework_TestCase
 {
